Fix Codex review suggestions for tasks, persistence, and navigation

+++ app/src/main/java/com/example/brainnote/feature/home/HomeScreen.kt
@@ -45,7 +45,7 @@ import androidx.core.view.WindowInsetsControllerCompat
 @Composable
 fun HomeScreen(
     onAddNoteClick: () -> Unit = {},
-    onTaskCardClick: () -> Unit = {}
+    onTaskCardClick: (Int) -> Unit = {}
 ) {
     var selectedTab by remember { mutableStateOf(0) }
     val primaryColor = Color(0xFF7445C8)
